Start using the new package manifest validator
Start using the newly introduced system for package manifest validation,
serialization, and deserialization
