+++ django/repository/models.py
@@ -343,6 +343,8 @@ class PackageVersion(models.Model):
     name = models.CharField(
         max_length=Package._meta.get_field("name").max_length,
     )
+
+    # TODO: Split to three fields for each number in the version for better querying performance
     version_number = models.CharField(
         max_length=16,
     )
